[#42] Add acl feature flag and deactive it by default
