+++ iceoryx2-bb/posix/Cargo.toml
@@ -10,6 +10,9 @@ repository = { workspace = true }
 rust-version = { workspace = true }
 version = { workspace = true }
 
+[features]
+acl = ["iceoryx2-pal-posix/acl"]
+
 [dependencies]
 iceoryx2-bb-container = { workspace = true }
 iceoryx2-bb-system-types = { workspace = true }
