+++ iceoryx2-bb/posix/src/lib.rs
@@ -12,6 +12,7 @@
 
 //! Abstraction of POSIX constructs with a safe API
 
+#[cfg(feature = "acl")]
 use access_control_list::AccessControlListError;
 use barrier::BarrierCreationError;
 use clock::ClockError;
@@ -31,6 +32,7 @@ use thread::ThreadError;
 use unix_datagram_socket::UnixDatagramError;
 use user::UserError;
 
+#[cfg(feature = "acl")]
 pub mod access_control_list;
 pub mod access_mode;
 pub mod adaptive_wait;
@@ -70,9 +72,31 @@ pub mod unix_datagram_socket;
 pub mod unmovable_ipc_handle;
 pub mod user;
 
+#[cfg(feature = "acl")]
 enum_gen! {Error
   generalization:
     AccessControlList <= AccessControlListError,
+    Barrier <= BarrierCreationError,
+    Clock <= ClockError,
+    Directory <= DirectoryError,
+    File <= FileError,
+    FileLock <= FileLockError,
+    Group <= GroupError,
+    MemoryLock <= MemoryLockError,
+    Mutex <= MutexError,
+    Process <= ProcessError,
+    ReadWriteMutex <= ReadWriteMutexError,
+    Semaphore <= SemaphoreError,
+    SharedMemory <= SharedMemoryCreationError,
+    Signal <= SignalError,
+    Thread <= ThreadError,
+    User <= UserError,
+    UnixDatagramSocket <= UnixDatagramError
+}
+
+#[cfg(not(feature = "acl"))]
+enum_gen! {Error
+  generalization:
     Barrier <= BarrierCreationError,
     Clock <= ClockError,
     Directory <= DirectoryError,
