+++ iceoryx2-pal/posix/Cargo.toml
@@ -14,6 +14,9 @@ version = { workspace = true }
 cc = { workspace = true }
 bindgen = { workspace = true }
 
+[features]
+acl = []
+
 [dependencies]
 iceoryx2-pal-concurrency-sync = { workspace = true }
 iceoryx2-pal-configuration = { workspace = true }
